Added progress bar widget. Refactored a little bit of code. Added better widget descriptions in the documentation.

+++ source/tkd/image/image.d
@@ -130,7 +130,7 @@ class Image : Element
 	 * Params:
 	 *     gamma = The destination gamma.
 	 */
-	public void setGamma(float gamma)
+	public void setGamma(double gamma)
 	{
 		if (gamma < 0)
 		{
